requirements.txt
# Money Matrix - Python Dependencies
# All dependencies are free and open-source

# Core Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Firebase Authentication
firebase-admin==6.3.0

# Database ORM
SQLAlchemy==2.0.23

# Data Validation & Serialization
marshmallow==3.20.1

# Data Analysis & Analytics
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# CORS Support
Flask-CORS==4.0.0

# Environment Variables
python-dotenv==1.0.0

# Production Server (optional)
gunicorn==21.2.0

# Development Tools (optional)
pytest==7.4.3
pytest-flask==1.3.0
